requirements/base.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0
albumentations>=1.4.0
scikit-image>=0.22.0
scipy>=1.11.0
timm>=0.9.0
pyyaml>=6.0
loguru>=0.7.0
rich>=13.7.0
tqdm>=4.66.0
